Move publish_as_nanopub() to RdfWrapper. No longer a method of fairstep or fairworkflow.
